Fix detection of Backspace and Delay elements
Previously, during `extractTextFromElement` and `cloneElementWithSpecifiedTextAtIndex`,
the code would try to determine if an element it was visiting was a
`Backspace` or a `Delay` element by checking `element.type.name`. This
worked fine most of the time, but if the code was minified,
`element.type.name` was no longer `Backspace` or `Delay` because the
name of those Components had been changed after minification.

In order to fix this, check the reference to the type (`element.type`) instead of
checking `element.type.name`